requirements.txt
Django==2.2.7
django-sendsms==0.4
pytz==2019.3
sqlparse==0.3.0
